+++ src/lib/escape/index.ts
@@ -1,1 +1,0 @@
-export * from './escape-html-entities.js';