+++ src/lib/index.ts
@@ -1,6 +1,0 @@
-export * from './compose.js';
-export * from './error.js';
-export * from './escape/index.js';
-export * from './filter/index.js';
-export * from './sort/index.js';
-export * from './util/index.js';