+++ src/util/index.ts
@@ -1,3 +1,2 @@
 export * from './convert-number-to-string.js';
 export * from './is-stringable.js';
-export * from './is-valid-attribute-name.js';