+++ tests/functional/util/convert-number-to-string.js
@@ -2,7 +2,7 @@ import { test } from 'uvu';
 import * as assert from 'uvu/assert';
 import {
     convertNumberToString
-} from '@mcaskill/html-build-attributes/lib/util/convert-number-to-string.js';
+} from '@mcaskill/html-build-attributes/util/convert-number-to-string.js';
 
 test('should return empty string when value is not a finite number', () => {
     [